feat(frontend): move scheduler to lib
