+++ src/frontend/src/icp/schedulers/wallet.scheduler.ts
@@ -2,6 +2,12 @@ import { WALLET_TIMER_INTERVAL_MILLIS } from '$icp/constants/ic.constants';
 import type { IcTransactionAddOnsInfo, IcTransactionUi } from '$icp/types/ic';
 import type { GetTransactions } from '$icp/types/ic.post-message';
 import { queryAndUpdate } from '$lib/actors/query.ic';
+import {
+	SchedulerTimer,
+	type Scheduler,
+	type SchedulerJobData,
+	type SchedulerJobParams
+} from '$lib/schedulers/scheduler';
 import type {
 	PostMessageDataResponseError,
 	PostMessageDataResponseWallet,
@@ -11,12 +17,6 @@ import type { CertifiedData } from '$lib/types/store';
 import type { Transaction, TransactionWithId } from '@dfinity/ledger-icp';
 import type { IcrcTransaction, IcrcTransactionWithId } from '@dfinity/ledger-icrc';
 import { isNullish, jsonReplacer } from '@dfinity/utils';
-import {
-	SchedulerTimer,
-	type Scheduler,
-	type SchedulerJobData,
-	type SchedulerJobParams
-} from './scheduler';
 
 type IndexedTransaction<T> = T & IcTransactionAddOnsInfo;
